src/pytorch_sac/requirements_dmc.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dm-control==1.0.16
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.3.0
filelock==3.13.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
glfw==2.7.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
importlib_resources==6.3.2
Jinja2==3.1.3
labmaze==1.0.6
lxml==5.1.0
MarkupSafe==2.1.5
mpmath==1.3.0
mujoco==3.1.3
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
PyOpenGL==3.1.7
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.31.0
scipy==1.10.1
sentry-sdk==1.42.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
termcolor==2.4.0
torch==2.2.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
wandb==0.16.4
zipp==3.18.1
